Renamed utils.py to general.py
