+++ src/highrl/utils/teacher_utils.py
@@ -4,7 +4,7 @@ import math
 import logging
 from prettytable import PrettyTable
 
-from highrl.utils.utils import TeacherConfigs, Position
+from highrl.utils.general import TeacherConfigs, Position
 from highrl.utils.training_utils import TeacherMetrics, RobotMetrics
 from highrl.utils.teacher_checker import convex_hull_difficulty
 from highrl.obstacle import SingleObstacle
@@ -129,7 +129,6 @@ def get_robot_position_from_action(
         action_table.add_column(fieldname=names[idx], column=[action_val])
 
     _LOG.info("====== Teacher action for Session %i ========", opt.time_steps)
-    _LOG.info(action_table)
     robot_pos = Position()
     goal_pos = Position()
     robot_pos.x_pos = min(int(opt.width * planner_output["robot_x"]), opt.width - 2)
